Move active set prob to Roles & Performance card
